+++ src/settings/MappingProfiles/ViewMappingProfile/ViewMappingProfile.js
@@ -61,6 +61,7 @@ import {
   marcFieldProtectionSettingsShape,
   showActionMenu,
   isMARCType,
+  BASE_URLS,
 } from '../../../utils';
 
 import sharedCss from '../../../shared.css';
@@ -176,7 +177,7 @@ export class ViewMappingProfile extends Component {
         entity={this}
         menu={menu}
         recordId={record?.id}
-        baseUrl="/settings/data-import/mapping-profiles"
+        baseUrl={BASE_URLS.MAPPING_PROFILE}
       />
     );
   };
@@ -282,6 +283,7 @@ export class ViewMappingProfile extends Component {
         history={history}
         location={location}
         recordId={id}
+        baseUrl={BASE_URLS.MAPPING_PROFILE}
       >
         <FullScreenView
           data-test-mapping-profile-details